// types.ts
export interface User {
  name: string;
  email: string;
  picture: string;
}
// ...
export interface ChatMessage {
  id:string;
  user: User;
  text: string;
  timestamp: string;
}
// ...
export interface SyntheticUser {
  id: string;
  name: string;
  avatarUrl: string;
  personaSummary: string;
}